feat: update monitor dashboard
