+++ src/components/DashboardCard/index.less
@@ -1,6 +1,6 @@
 .dashboard-card {
 	padding: 5px;
-  min-height: 293px;
+  // min-height: 293px;
   height: 320px;
 
 	> .inner {
