+++ src/pages/MachineDashboard/Cards/DiskCard.tsx
@@ -34,11 +34,11 @@ const DiskCard = forwardRef((props: IProps, ref) => {
     const queries: any = [
       {
         refId: 'diskSize',
-        query: `node_filesystem_size_bytes{fstype=~"ext.*|xfs",mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}}`
+        query: `node_filesystem_size_bytes{mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}}`
       },
       {
         refId: 'diskUsed',
-        query: `node_filesystem_size_bytes{fstype=~"ext.*|xfs",mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}} - node_filesystem_free_bytes{fstype=~"ext.*|xfs",mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}}`
+        query: `node_filesystem_size_bytes{mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}} - node_filesystem_free_bytes{fstype=~"ext.*|xfs",mountpoint !~".*pod.*"${clusterSuffix1}${instanceSuffix}}`
       }
     ];
     const data = await asyncBatchQueries(queries);
